requirements.txt
Flask==2.2.3
gunicorn==20.1.0
werkzeug==2.2.3
python-dotenv
psycopg2-binary